Extract useFormField hook
