+++ projects/Mallard/src/screens/auth-switcher-screen.tsx
@@ -13,8 +13,9 @@ import { useModal } from 'src/components/modal'
 import { SubNotFoundModalCard } from 'src/components/sub-not-found-modal-card'
 import { routeNames } from 'src/navigation'
 import { SubFoundModalCard } from 'src/components/sub-found-modal-card'
-import { LogIn, useFormField } from './log-in'
+import { LogIn } from './log-in'
 import isEmail from 'validator/lib/isEmail'
+import { useFormField } from 'src/hooks/use-form-field'
 
 const useRandomState = () =>
     useState(
