+++ projects/Mallard/src/screens/log-in.tsx
@@ -14,6 +14,7 @@ import { useInsets } from 'src/hooks/use-insets'
 import { color } from 'src/theme/color'
 import { Link } from 'src/components/link'
 import { getFont } from 'src/theme/typography'
+import { FormField } from 'src/hooks/use-form-field'
 
 const SocialButton = ({
     children,
@@ -41,39 +42,6 @@ const SocialButton = ({
         {children}
     </Button>
 )
-
-interface FormField {
-    value: string
-    setValue: (value: string) => void
-    error: string | null
-}
-
-const useFormField = (
-    initialValue: string,
-    {
-        validator,
-        onSet,
-    }: {
-        validator: (value: string) => string | null
-        onSet?: (value: string) => void
-    },
-): FormField => {
-    const [value, setValue] = useState(initialValue)
-    const [error, setError] = useState<string | null>(null)
-
-    useEffect(() => {
-        setError(validator(value))
-    }, [validator, value])
-
-    return {
-        value,
-        setValue: value => {
-            onSet && onSet(value)
-            setValue(value)
-        },
-        error,
-    }
-}
 
 const Shadow = () => (
     <View
@@ -438,4 +406,4 @@ const LogIn = ({
     )
 }
 
-export { LogIn, useFormField }
+export { LogIn }
